Add v2.0: Support planar extensional flow
This version adds support for planar extensional flow in addition to the
prior support of pure simple shear. While the simple shear was based on
the paper from Doi and Edwards (1978) and the paper from Talbot et al.
(2024), the extensional operator is newly analytically solved. More
details are provided in the documentation pdf file (see `doc/`).

The planar extensional flow introduces a new control parameter: the
extensional rate, and thus a second Peclet number. All functions have
been updated to now take both parameters. Both shear and extensional
flow may be combined.

The code was validated by checking the assembled matrix with the exact
integral values and with a separate implementation in Dedalus, a
spectral library for python. There integrals are handled within the
library. Thus the code should be correct, unless the initial
equations were wrong.