Split matrices and (steady) avoid recomputations
To help avoid re-computations of the matrices, which is time consuming,
the build matrix function now only provides the matrices independent of
simulation parameters.

For steady state, the `fp_steady` function now computes the matrices
only once. In addition, the automatic `Lmax` selection is removed in
favour of an adaptive method. This reuses the pre-computed large
matrices using sub-blocks and compares a high-resolution solution with
a low-resolution solution to estimate the error in `b_{2,m}`. Due to the
increased number of linear systems to solve, this is slower than fixing
`Lmax` manually but avoids making this choice.

Note that extension flow is significantly slower than shear flow because
it requires a much larger `Lmax`.