v3.0: Update documentation for finalizing v3.0
v3.0 pre-computes matrices to avoid unnecessary and slow re-calcuations
of the matrices. Note that this can be made still more efficient by
storing a large version of the matrices in a file and reading when
necessary. However, for now this is not planned.

Some other changes were also made, relating to plotting and file names.
See commits since v2.0 for all changes.