Store and load matrices if possible
Still much of the computational time is invested in computing the
matrices. However, these never change unless Lmax is increased or
decreased. As such, they are now automatically stored and loaded when
calling the build matrix function. If no file is available, they are
calculated and saved in a file under 'data/'.