Add option for optional matrix store
The matrices can now optionally be stored when calling the build
function rather than always stored. This is set to false for the helper
functions 'solve_steady' and 'solve_unsteady'.

The full simulations are hard coded to store them, as these are intended
to run multiple shear rates and extension rates.