Fix: xz and xy shear instead of incorrect zx or yx
A minus sign in the 'fp_init' and 'fp_steady' resulted in shear being
solved for zx instead of xz. This also cased a 90 degree shift for the
tranformed xy plane, giving yx shear insteady of xy. The minus sign is
now removed and the shear should be in the correct direction.